+++ embedded/cult_connect_embbeded/src/main.cpp
@@ -160,7 +160,7 @@ void loop()
         }
 
         // Check if sensors measure ticker is up
-        for (int i = 0; i < moduleConfig.getNbSensors(); i++)
+        for (uint8_t i = 0; i < moduleConfig.getNbSensors(); i++)
         {
             moduleConfig.sensors[i]->update();
         }
